fix initialize temp.json on server

+++ server/api.py
@@ -158,6 +158,7 @@ def post(teks):
         return response
 
 if __name__ == "__main__":
-    with open('temp.json', 'w') as output:
-        json.dump({"sekolah": "", "start": False}, output)
+    if os.path.isfile('temp.json') == False:
+        with open('temp.json', 'w') as output:
+            json.dump({"sekolah": "", "start": False}, output)
     socketIo.run(app, host='0.0.0.0', port=5000)
